+++ strato/VM/SolidVM/solid-vm-parser/tests/ParserSpec.hs
@@ -13,11 +13,6 @@ import SolidVM.Model.CodeCollection.Statement
 import SolidVM.Model.Type
 import SolidVM.Solidity.Parse.Lexer
 import SolidVM.Solidity.Parse.Statement
-import qualified Data.Map as M
--- import SolidVM.Solidity.Parse.Declarations
--- import SolidVM.Model.CodeCollection.Def as Def
-
--- import SolidVM.Solidity.Parse.UnParser
 import SolidVM.Solidity.Parse.ParserTypes 
 
 import           Data.Source.Annotation as SA 
@@ -50,7 +45,7 @@ dummyAnnotation =
 spec :: Spec
 spec = do
   describe "String lexing" $ do
-    let parseStr = runParser (stringLiteral <* eof) (ParserState "" "" M.empty) ""
+    let parseStr = runParser (stringLiteral <* eof) initialParserState ""
         cases = [ ([r|"ok"|], "ok")
                 , ([r|"ok" |], "ok")
                 ]
@@ -58,7 +53,7 @@ spec = do
       it ("can parse " ++ show input) $ parseStr input `shouldBe` Right want
 
   describe "Expression parsing" $ do
-    let parseExpr = fmap (fmap (const ())) . runParser expression (ParserState "" "" M.empty) ""
+    let parseExpr = fmap (fmap (const ())) . runParser expression initialParserState ""
         cases = [ ("x++", PlusPlus () (Variable () "x"))
                 , ("++x", Unitary () "++" (Variable () "x"))
                 , ("--x", Unitary () "--" (Variable () "x"))
@@ -142,7 +137,7 @@ spec = do
 -}
 
   describe "Statement parsing" $ do
-    let parseStatement = fmap (fmap (const ())) . runParser statement (ParserState "" "" M.empty) ""
+    let parseStatement = fmap (fmap (const ())) . runParser statement initialParserState ""
         scases = [ ("x++;", SimpleStatement $ ExpressionStatement $ PlusPlus () $ Variable () "x")
                  , ("assembly { dst := mload(add(src, 32)) }",
                       AssemblyStatement $ MloadAdd32 "dst" "src")