+++ strato/api/core/src/Handlers/AccountInfo.hs
@@ -29,7 +29,7 @@ import           Servant
 import           Servant.Client
 --import           Servant.Swagger.Tags
 
-
+import           Blockchain.Data.CirrusDefs
 import           Blockchain.Data.DataDefs
 import           Blockchain.Data.Json
 import           Blockchain.DB.SQLDB
@@ -254,3 +254,10 @@ getCodeByteStringFromPostgres cHash =
         E.from $ \(codeRef) -> do
         E.where_ (codeRef E.^. CodeRefCodeHash E.==. E.val cHash)
         return codeRef
+
+getX509CertForAccount :: HasCirrus m => Address -> m (Maybe Certificate)
+getX509CertForAccount addr = do
+  fmap (listToMaybe . map E.entityVal) . cirrusQuery . E.select $
+    E.from $ \(certificate) -> do
+      E.where_ (certificate E.^. CertificateUserAddress E.==. E.val addr)
+      return $ certificate