+++ strato/indexer/slipstream/exec_src/Main.hs
@@ -13,6 +13,7 @@ import Control.Concurrent
 import Control.Monad
 import Control.Monad.IO.Class
 import Control.Monad.Trans.Resource
+import Control.Monad.Trans.Reader
 import qualified Data.ByteString as B
 import qualified Data.ByteString.Char8 as BC
 import Data.String
@@ -38,7 +39,7 @@ import Slipstream.Globals
 import Slipstream.GlobalsColdStorage
 import Slipstream.Options
 import Slipstream.OutputData
-import Slipstream.Processor
+-- import Slipstream.Processor
 
 import SelectAccessible ()
 
@@ -57,6 +58,8 @@ createBlocEnv = liftIO $ do
                  , gasLimit=0
                  , globalNonceCounter=error("globalNonceCounter shouldn't be needed in slipstream, it is undefined")
                  , txTBQueue=error("txTBQueue shouldn't be needed in slipstream, it is undefined")
+                 , userRegistryAddress=0x0
+                 , useWalletsByDefault = error "useWalletsByDefault shouldn't be needed in slipstream"
     }
 
 
@@ -71,7 +74,6 @@ main = do
   runLoggingT
     . runResourceT
     . runKafkaM ("slipstream" :: KafkaClientId) (fromString flags_kafkahost, fromIntegral flags_kafkaport)
-    . runSQLM
     . withPostgresqlConn workerConnStr $ \workerConn -> do
 
     $logInfoS "main" "Welcome to Slipstream!!!!"
@@ -93,10 +95,6 @@ main = do
 
     handle <- runSqlConn initStorage workerConn
     gref <- newGlobals handle (CirrusHandle conn S.empty)
-
-    --Create Mercata Abstract tables
-    generateAssetTable conn gref
-    generateSaleTable conn gref
-    generateUserTable conn gref
     
-    getAndProcessMessages env conn gref
+    flip runReaderT gref . runSQLM $
+      getAndProcessMessages env conn